requirements.txt
streamlit
torch
torchaudio
chatterbox-tts
streamlit-audiorecorder
